+++ src/Checkout/SPBCheckout/Service/SPBCheckoutDataService.php
@@ -17,6 +17,12 @@ use Symfony\Component\Routing\RouterInterface;
 
 class SPBCheckoutDataService
 {
+    private const APM_BLIK = 'blik';
+    private const APM_EPS = 'eps';
+    private const APM_GIROPAY = 'giropay';
+    private const APM_P24 = 'p24';
+    private const APM_SOFORT = 'sofort';
+
     /**
      * @var PaymentMethodUtil
      */
@@ -75,6 +81,30 @@ class SPBCheckoutDataService
         }
 
         return $spbCheckoutButtonData;
+    }
+
+    /**
+     * @return string[]
+     */
+    public function getDisabledAlternativePaymentMethods(float $totalPrice, string $currencyIsoCode): array
+    {
+        $disabled = [];
+
+        if ($totalPrice < 1.0 && $currencyIsoCode === 'EUR') {
+            $disabled[] = self::APM_EPS;
+            $disabled[] = self::APM_GIROPAY;
+            $disabled[] = self::APM_SOFORT;
+        }
+
+        if ($totalPrice < 1.0 && $currencyIsoCode === 'PLN') {
+            $disabled[] = self::APM_BLIK;
+        }
+
+        if (($totalPrice < 1.0 || $totalPrice > 55000.0) && $currencyIsoCode === 'PLN') {
+            $disabled[] = self::APM_P24;
+        }
+
+        return $disabled;
     }
 
     private function getButtonLanguage(SwagPayPalSettingStruct $settings, SalesChannelContext $context): string