+++ src/Resources/app/storefront/src/main.js
@@ -7,8 +7,28 @@ import SwagPayPalInstallmentBanner from './installment/swag-paypal.installment-b
 
 // Register them via the existing PluginManager
 const PluginManager = window.PluginManager;
-PluginManager.register('SwagPayPalExpressButton', SwagPayPalExpressCheckoutButton, '[data-swag-paypal-express-button]');
-PluginManager.register('SwagPayPalSmartPaymentButtons', SwagPayPalSmartPaymentButtons, '[data-swag-paypal-smart-payment-buttons]');
-PluginManager.register('SwagPayPalMarks', SwagPayPalMarks, '[data-swag-paypal-marks]');
-PluginManager.register('SwagPayPalPlusPaymentWall', SwagPayPalPlusPaymentWall, '[data-swag-paypal-payment-wall]');
-PluginManager.register('SwagPayPalInstallmentBanner', SwagPayPalInstallmentBanner, '[data-swag-paypal-installment-banner]');
+PluginManager.register(
+    'SwagPayPalExpressButton',
+    SwagPayPalExpressCheckoutButton,
+    '[data-swag-paypal-express-button]'
+);
+PluginManager.register(
+    'SwagPayPalSmartPaymentButtons',
+    SwagPayPalSmartPaymentButtons,
+    '[data-swag-paypal-smart-payment-buttons]'
+);
+PluginManager.register(
+    'SwagPayPalMarks',
+    SwagPayPalMarks,
+    '[data-swag-paypal-marks]'
+);
+PluginManager.register(
+    'SwagPayPalPlusPaymentWall',
+    SwagPayPalPlusPaymentWall,
+    '[data-swag-paypal-payment-wall]'
+);
+PluginManager.register(
+    'SwagPayPalInstallmentBanner',
+    SwagPayPalInstallmentBanner,
+    '[data-swag-paypal-installment-banner]'
+);